[Core] Separate `PointObject` in different file
